+++ packages/sif-common-forms-ds/src/i18n/forms.messages.tsx
@@ -1,3 +1,4 @@
+import { FormattedMessage, IntlShape, useIntl } from 'react-intl';
 import { annetBarnMessages } from '../forms/annet-barn/i18n';
 import { bostedUtlandMessages } from '../forms/bosted-utland/i18n';
 import { enkeltdatoMessages } from '../forms/enkeltdatoer';
@@ -10,6 +11,9 @@ import { tidsperiodeMessages } from '../forms/tidsperiode/i18n';
 import { utenlandskNæringMessages } from '../forms/utenlandsk-næring/i18n';
 import { utenlandsoppholdMessages } from '../forms/utenlandsopphold/i18n';
 import { virksomhetMessages } from '../forms/virksomhet/i18n';
+import { typedIntlHelper } from '@navikt/sif-common-core-ds/src/utils/intlUtils';
+import { velgBarnFormPart_nb } from '../form-parts/velg-barn-form-part/i18n/nb';
+import { velgBarnFormPart_nn } from '../form-parts/velg-barn-form-part/i18n/nn';
 
 const nb = {
     ...annetBarnMessages.nb,
@@ -24,6 +28,7 @@ const nb = {
     ...virksomhetMessages.nb,
     ...medlemskapFormMessages.nb,
     ...enkeltdatoMessages.nb,
+    ...velgBarnFormPart_nb,
 };
 
 type FormsMessageKeys = keyof typeof nb;
@@ -41,6 +46,25 @@ const nn: Record<FormsMessageKeys, string> = {
     ...virksomhetMessages.nn,
     ...medlemskapFormMessages.nn,
     ...enkeltdatoMessages.nn,
+    ...velgBarnFormPart_nn,
+};
+
+interface FormsTextProps {
+    id: FormsMessageKeys;
+    values?: any;
+}
+
+export const getUiIntl = (intl: IntlShape) => {
+    return typedIntlHelper<keyof typeof nb>(intl);
+};
+
+export const useFormsIntl = () => {
+    const intl = useIntl();
+    return getUiIntl(intl);
+};
+
+export const FormsText = (props: FormsTextProps) => {
+    return <FormattedMessage {...props} />;
 };
 
 export const formsMessages = {