+++ packages/sif-common-forms-ds/src/index.ts
@@ -1,3 +1,4 @@
+export * from './form-parts/velg-barn-form-part/VelgBarnFormPart';
 export * from './forms/annet-barn';
 export * from './forms/bosted-utland';
 export * from './forms/ferieuttak';