+++ apps/omsorgspengesoknad-refactor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @navikt/omsorgspengesoknad
 
+## 1.39.9
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-forms-ds@20.0.0
+    - @navikt/sif-common-ui@0.10.0
+    - @navikt/sif-common-soknad-ds@25.0.0
+
 ## 1.39.8
 
 ### Patch Changes