+++ apps/opplaringspenger-soknad/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @navikt/opplaringspenger-soknad
 
+## 1.0.27
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-forms-ds@20.0.0
+    - @navikt/sif-common-ui@0.10.0
+    - @navikt/sif-common-soknad-ds@25.0.0
+
 ## 1.0.26
 
 ### Patch Changes