+++ apps/sif-demo-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @navikt/sif-demo-app
 
+## 0.4.26
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-forms-ds@20.0.0
+    - @navikt/sif-common-ui@0.10.0
+    - @navikt/sif-common-soknad-ds@25.0.0
+
 ## 0.4.25
 
 ### Patch Changes